refactor(max-zoom): rename everything from max-zoom to max-zoom-service
Breaking Changes
