+++ examples/max-zoom-service/README.md
