+++ examples/max-zoom-service/components/map-canvas/map-canvas.module.css
