+++ examples/max-zoom-service/components/max-zoom-service/max-zoom-service.tsx
