+++ examples/max-zoom-service/index.html
